LPS-59277 - Register command for EditLicenseAction action
