+++ portal-impl/src/META-INF/util-spring.xml
@@ -605,6 +605,8 @@
 	</bean>
 	<bean id="com.liferay.portlet.softwarecatalog.action.EditFrameworkVersionMVCActionCommand" class="com.liferay.portlet.softwarecatalog.action.EditFrameworkVersionMVCActionCommand" />
 	<bean id="com.liferay.portlet.softwarecatalog.action.EditFrameworkVersionMVCRenderCommand" class="com.liferay.portlet.softwarecatalog.action.EditFrameworkVersionMVCRenderCommand" />
+	<bean id="com.liferay.portlet.softwarecatalog.action.EditLicenseMVCActionCommand" class="com.liferay.portlet.softwarecatalog.action.EditLicenseMVCActionCommand" />
+	<bean id="com.liferay.portlet.softwarecatalog.action.EditLicenseMVCRenderCommand" class="com.liferay.portlet.softwarecatalog.action.EditLicenseMVCRenderCommand" />
 	<bean id="com.liferay.portlet.social.service.permission.SocialActivityPermissionUtil" class="com.liferay.portlet.social.service.permission.SocialActivityPermissionUtil">
 		<property name="socialActivityPermission">
 			<bean class="com.liferay.portlet.social.service.permission.SocialActivityPermissionImpl" />
