--- portal-impl/src/com/liferay/portlet/softwarecatalog/action/EditLicenseAction.java
@@ -1,140 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portlet.softwarecatalog.action;
-
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.struts.PortletAction;
-import com.liferay.portlet.softwarecatalog.LicenseNameException;
-import com.liferay.portlet.softwarecatalog.NoSuchLicenseException;
-import com.liferay.portlet.softwarecatalog.RequiredLicenseException;
-import com.liferay.portlet.softwarecatalog.service.SCLicenseServiceUtil;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.PortletConfig;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import org.apache.struts.action.ActionForm;
-import org.apache.struts.action.ActionForward;
-import org.apache.struts.action.ActionMapping;
-
-/**
- * @author Jorge Ferrer
- */
-public class EditLicenseAction extends PortletAction {
-
-	@Override
-	public void processAction(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse)
-		throws Exception {
-
-		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
-
-		try {
-			if (cmd.equals(Constants.ADD) || cmd.equals(Constants.UPDATE)) {
-				updateLicense(actionRequest);
-			}
-			else if (cmd.equals(Constants.DELETE)) {
-				deleteLicense(actionRequest);
-			}
-
-			sendRedirect(actionRequest, actionResponse);
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchLicenseException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-
-				setForward(actionRequest, "portlet.software_catalog.error");
-			}
-			else if (e instanceof LicenseNameException ||
-					 e instanceof RequiredLicenseException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-			}
-			else {
-				throw e;
-			}
-		}
-	}
-
-	@Override
-	public ActionForward render(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		try {
-			ActionUtil.getLicense(renderRequest);
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchLicenseException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(renderRequest, e.getClass());
-
-				return actionMapping.findForward(
-					"portlet.software_catalog.error");
-			}
-			else {
-				throw e;
-			}
-		}
-
-		return actionMapping.findForward(
-			getForward(renderRequest, "portlet.software_catalog.edit_license"));
-	}
-
-	protected void deleteLicense(ActionRequest actionRequest) throws Exception {
-		long licenseId = ParamUtil.getLong(actionRequest, "licenseId");
-
-		SCLicenseServiceUtil.deleteLicense(licenseId);
-	}
-
-	protected void updateLicense(ActionRequest actionRequest) throws Exception {
-		long licenseId = ParamUtil.getLong(actionRequest, "licenseId");
-
-		String name = ParamUtil.getString(actionRequest, "name");
-		String url = ParamUtil.getString(actionRequest, "url");
-		boolean openSource = ParamUtil.getBoolean(actionRequest, "openSource");
-		boolean active = ParamUtil.getBoolean(actionRequest, "active");
-		boolean recommended = ParamUtil.getBoolean(
-			actionRequest, "recommended");
-
-		if (licenseId <= 0) {
-
-			// Add license
-
-			SCLicenseServiceUtil.addLicense(
-				name, url, openSource, active, recommended);
-		}
-		else {
-
-			// Update license
-
-			SCLicenseServiceUtil.updateLicense(
-				licenseId, name, url, openSource, active, recommended);
-		}
-	}
-
-}
