+++ portal-web/docroot/WEB-INF/struts-config.xml
@@ -183,11 +183,6 @@
 
 		<!-- Software Catalog -->
 
-		<action path="/software_catalog/edit_license" type="com.liferay.portlet.softwarecatalog.action.EditLicenseAction">
-			<forward name="portlet.software_catalog.edit_license" path="portlet.software_catalog.edit_license" />
-			<forward name="portlet.software_catalog.error" path="portlet.software_catalog.error" />
-		</action>
-
 		<action path="/software_catalog/edit_product_entry" type="com.liferay.portlet.softwarecatalog.action.EditProductEntryAction">
 			<forward name="portlet.software_catalog.edit_product_entry" path="portlet.software_catalog.edit_product_entry" />
 			<forward name="portlet.software_catalog.error" path="portlet.software_catalog.error" />
