+++ portal-web/docroot/html/portlet/software_catalog/edit_license.jsp
@@ -24,7 +24,7 @@ SCLicense license = (SCLicense)request.getAttribute(WebKeys.SOFTWARE_CATALOG_LIC
 long licenseId = BeanParamUtil.getLong(license, request, "licenseId");
 %>
 
-<form action="<portlet:actionURL><portlet:param name="struts_action" value="/software_catalog/edit_license" /></portlet:actionURL>" method="post" name="<portlet:namespace />fm" onSubmit="<portlet:namespace />saveLicense(); return false;">
+<form action="<portlet:actionURL name="/software_catalog/edit_license" var="" />" method="post" name="<portlet:namespace />fm" onSubmit="<portlet:namespace />saveLicense(); return false;">
 <input name="<portlet:namespace /><%= Constants.CMD %>" type="hidden" value="" />
 <input name="<portlet:namespace />redirect" type="hidden" value="<%= HtmlUtil.escapeAttribute(redirect) %>" />
 <input name="<portlet:namespace />licenseId" type="hidden" value="<%= licenseId %>" />
